pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exocort"
version = "0.1.0"
description = "Exocort collector, capture agents, and processor"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "python-dotenv",
  "pydantic",
  "requests",
  "fastapi",
  "python-multipart",
  "uvicorn[standard]",
  "sounddevice",
  "webrtcvad-wheels",
  "mss",
  "opencv-python==4.10.0.84",
  "numpy",
]

[project.optional-dependencies]
collector = [
  "fastapi",
  "uvicorn[standard]",
  "python-multipart",
]
audio = [
  "sounddevice",
  "webrtcvad-wheels",
]
screen = [
  "mss",
  "opencv-python",
  "numpy",
  "pywinctl",
]
processor = [
]
test = [
  "pytest",
  "pytest-cov",
]

[project.scripts]
exocort = "runner:main"
exocort-collector = "collector.app:main"
exocort-audio = "capture.audio:main"
exocort-screen = "capture.screen:main"
exocort-processor = "processor.cli:main"

[tool.setuptools]
package-dir = {"" = "exocort"}
py-modules = ["runner"]

[tool.setuptools.packages.find]
where = ["exocort"]

[tool.pytest.ini_options]
pythonpath = ["exocort", "."]
testpaths = ["tests", "services"]
markers = [
  "unit: fast unit tests",
  "integration: internal integration tests",
  "service: service-level tests",
  "slow: slower tests",
  "llm: llm-focused tests",
  "ocr: ocr-focused tests",
  "stt: stt-focused tests",
]
addopts = "-ra"
